pyproject.toml
[project]
name = "artificial-architect"
version = "0.1.0"
description = "Autonomous Agent for High-Stakes Software Engineering Evaluation"
authors = [
    {name = "Antigravity", email = "agent@google.com"},
]
dependencies = [
    "langchain>=0.1.0",
    "langgraph>=0.0.1",
    "langchain-anthropic",
    "langchain-openai",
    "pydantic>=2.0.0",
    "neo4j>=5.0.0",
    "docker>=7.0.0",
    "rich>=13.0.0",  # For pretty console output
    "python-dotenv>=1.0.0",
    "requests>=2.31.0"
]
requires-python = ">=3.10"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
